Set the navigation generator service to "public"
I was attempting to use this in a Sculpin 3.0.0-rc2 project and the service not being public seemed to prevent the service container from properly adding it to the event subscriber. There might be a way for Sculpin 3.x to fix this potential BC break for all event listener based extensions, but in the meantime I thought a PR might be a good idea.